backend/requirements.txt
fastapi
uvicorn[standard]
numpy
python-multipart
cadquery
cadquery-ocp
requests
python-dotenv
reportlab
